+++ src/main/java/com/cloudogu/scm/review/pullrequest/dto/PullRequestDto.java
@@ -44,6 +44,7 @@ import java.util.Set;
 public class PullRequestDto extends HalRepresentation {
   private String id;
   private DisplayedUserDto author;
+  private DisplayedUserDto reviser;
   @NotBlank
   private String source;
   @NotBlank