+++ src/test/java/com/cloudogu/scm/review/pullrequest/service/DefaultPullRequestServiceTest.java
@@ -76,6 +76,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
@@ -428,6 +429,11 @@ class DefaultPullRequestServiceTest {
 
     @Test
     void shouldSetPullRequestRejectedAndSendEvent() {
+      Subject subject = mock(Subject.class, RETURNS_DEEP_STUBS);
+      shiroRule.setSubject(subject);
+      User currentUser = new User("currentUser");
+      when(subject.getPrincipals().oneByType(User.class)).thenReturn(currentUser);
+
       doReturn(Branch.normalBranch("source", "1"))
         .when(branchResolver).resolve(REPOSITORY, pullRequest.getSource());
       doReturn(Branch.normalBranch("target", "2"))
@@ -436,6 +442,7 @@ class DefaultPullRequestServiceTest {
 
       service.setRejected(REPOSITORY, pullRequest.getId(), REJECTED_BY_USER);
 
+      assertThat(pullRequest.getReviser()).isEqualTo(currentUser.getName());
       assertThat(pullRequest.getStatus()).isEqualTo(REJECTED);
       assertThat(pullRequest.getSourceRevision()).isEqualTo("1");
       assertThat(pullRequest.getTargetRevision()).isEqualTo("2");
@@ -468,11 +475,17 @@ class DefaultPullRequestServiceTest {
 
     @Test
     void shouldSetPullRequestMergedAndSendEvent() {
+      Subject subject = mock(Subject.class, RETURNS_DEEP_STUBS);
+      shiroRule.setSubject(subject);
+      User currentUser = new User("currentUser");
+      when(subject.getPrincipals().oneByType(User.class)).thenReturn(currentUser);
+
       pullRequest.setStatus(OPEN);
 
       service.setMerged(REPOSITORY, pullRequest.getId());
 
       assertThat(pullRequest.getStatus()).isEqualTo(MERGED);
+      assertThat(pullRequest.getReviser()).isEqualTo(currentUser.getName());
       assertThat(eventCaptor.getAllValues()).hasSize(1);
       PullRequestMergedEvent event = (PullRequestMergedEvent) eventCaptor.getValue();
       assertThat(event.getRepository()).isSameAs(REPOSITORY);
@@ -592,6 +605,6 @@ class DefaultPullRequestServiceTest {
   }
 
   private PullRequest createPullRequest(String id, Instant creationDate, Instant lastModified) {
-    return new PullRequest(id, "source", "target", "pr", "description", null, creationDate, lastModified, OPEN, emptySet(), new HashMap<>(), "", "", emptySet(), null, false, Collections.emptyList());
+    return new PullRequest(id, "source", "target", "pr", "description", null, null, creationDate, lastModified, OPEN, emptySet(), new HashMap<>(), "", "", emptySet(), null, false, Collections.emptyList());
   }
 }