+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Show pull request reviser for merged/rejected pull requests
+
 ## 2.8.0 - 2021-04-07
 ### Added
 - Enable anchor links for pull request comments ([#127](https://github.com/scm-manager/scm-review-plugin/pull/127))